Release 3.0.1 - Release 3.0.1 - fixed Embedding
- Changed return type of getDimension method from long to int
- Updated version number in pom.xml and README.md to 3.0.1

+++ src/main/java/de/MCmoderSD/openai/prompts/EmbeddingPrompt.java
@@ -83,7 +83,7 @@ public class EmbeddingPrompt {
         return user;
     }
 
-    public long getDimension() {
+    public int getDimension() {
         return dimension;
     }
